refactor tasks refactor model
