+++ .gitignore
@@ -97,3 +97,4 @@ ENV/
 # jetbrains products
 .idea/
 .pytest_cache/
+config.yaml
