+++ config-template.yaml
@@ -1,5 +1,11 @@
 default:
-  DB_CONNECTION: mongodb://localhost/database_name
+
+  AIRTABLE_API_ENDPOINT: <your airtable database endpoint>
+  AIRTABLE_API_KEY: <your airtable api key>
+
+  AIRTABLE_EXPT_TABLE: 'Experiments'
+  AIRTABLE_SAMPLE_TABLE: 'Samples'
+  AIRTABLE_TASK_TABLE: 'Tasks'
 
   RAW_BUCKET: s3://<your raw data bucket>
   S3_BUCKET: s3://<your processed data bucket>
