+++ ob_pipelines/__init__.py
@@ -4,12 +4,11 @@ import traceback
 from datetime import datetime
 
 import luigi
-from mongoengine import connect
 
 import ob_pipelines
-from ob_pipelines.batch import JobTask
-from ob_pipelines.config import settings
-from ob_pipelines.entities import Job, Task
+from ob_pipelines.batch import TaskWrapper
+from ob_pipelines.entities.persistence import create_task, get_task_by_key, update_task
+from ob_pipelines.entities.task import Task
 
 logger = logging.getLogger(__name__)
 
@@ -22,27 +21,28 @@ ch.setLevel(os.environ.get('LOGGING_LEVEL') or logging.DEBUG)
 ch.setFormatter(formatter)
 logger.addHandler(ch)
 
-connect(host=settings.db_connection)
+
+@TaskWrapper.event_handler(luigi.event.Event.START)
+def luigi_task_start(luigi_task: TaskWrapper):
+    task: Task = Task()
+    task.name = luigi_task.task_id_str
+    task.status = 'running'
+    task.started_at = datetime.utcnow()
+    luigi_task.task_id = create_task(task).key
 
 
-@JobTask.event_handler(luigi.event.Event.START)
-def luigi_task_start(task: JobTask):
-    job = Job.objects.get(id=task.job_id)
-    db_task: Task = Task(job=job, name=task.get_task_family(), started_at=datetime.now())
-    db_task.save()
-    task.db_task_id = db_task.id
+@TaskWrapper.event_handler(luigi.event.Event.SUCCESS)
+def luigi_task_success(luigi_task: TaskWrapper):
+    task = get_task_by_key(luigi_task.task_id)
+    task.completed_at = datetime.utcnow()
+    task.status = 'completed'
+    update_task(task)
 
 
-@JobTask.event_handler(luigi.event.Event.SUCCESS)
-def luigi_task_success(task: JobTask):
-    db_task: Task = Task.objects.get(id=task.db_task_id)
-    db_task.completed_at = datetime.now()
-    db_task.save()
-
-
-@JobTask.event_handler(luigi.event.Event.FAILURE)
-def luigi_task_failure(task: JobTask, exc):
-    db_task: Task = Task.objects.get(id=task.db_task_id)
-    db_task.completed_at = datetime.now()
-    db_task.exception = traceback.format_exc()
-    db_task.save()
+@TaskWrapper.event_handler(luigi.event.Event.FAILURE)
+def luigi_task_failure(luigi_task: TaskWrapper):
+    task = get_task_by_key(luigi_task.task_id)
+    task.completed_at = datetime.utcnow()
+    task.status = 'failed'
+    task.exception = traceback.format_exc()
+    update_task(task)
