+++ ob_pipelines/pipelines/xenograft/tasks/fastqc_trimmed.py
@@ -1,0 +1,8 @@
+from ob_pipelines.tasks.fastqc import FastQC
+from ob_pipelines.tasks.skewer import Skewer
+
+
+class FastQCTrimmed(FastQC):
+
+    def requires(self):
+        return Skewer(sample_id=self.sample_id)
