+++ ob_pipelines/tasks/ercc_quant.py
@@ -1,0 +1,36 @@
+from os import path as op
+
+from luigi.contrib.s3 import S3Target
+
+from ob_pipelines.batch import BatchTask
+from ob_pipelines.config import cfg
+from ob_pipelines.entities.sample import Sample
+from ob_pipelines.tasks.sample_fastq import SampleFastQ
+
+
+class ERCCQuant(BatchTask, Sample):
+    job_definition = 'kallisto'
+
+    @property
+    def parameters(self):
+        fq1, fq2 = self.input()
+        return {
+            'threads': '20',
+            'index': '/reference/kallisto/ercc.idx',
+            'reads1': fq1.path,
+            'reads2': fq2.path,
+            'strand_flag': '--rf-stranded',
+            'output': op.dirname(self.output()['abundance'].path) + '/'
+        }
+
+    def requires(self):
+        return SampleFastQ(sample_id=self.sample_id)
+
+    def output(self):
+        output_files = {
+            'abundance': 'abundance.tsv',
+            'h5': 'abundance.h5',
+            'run_info': 'run_info.json'
+        }
+        return {k: S3Target('{}/{}/ercc-rev-strand/{}'.format(cfg['S3_BUCKET'], self.sample_folder, fname))
+                for k, fname in output_files.items()}
